requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distro==1.8.0
Django==4.2.6
django-cors-headers==4.3.0
django-decouple==2.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==3.4
kombu==5.3.2
Markdown==3.5
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.39
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.0.7
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.6.0
